\name{bpNet-package}
\alias{bpNet-package}
\alias{bpNet}
\docType{package}
\title{
\packageTitle{bpNet}
}
\description{
\packageDescription{bpNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bpNet}
\packageIndices{bpNet}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{bpNet}

Maintainer: \packageMaintainer{bpNet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
